# Auction/auctionSwagger.yaml
openapi: '3.0.0'
info:
  version: 3.0.0
  title: Auction API's
  description: A sample API
  contact:
      name: API Support
      email: sunil.m@yash.com
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://{base_url}/{env_id}
    variables:
      base_url:
        default: localhost:8080
        description: server
      env_id:
        default: api
        description: path selector
paths:
  /auctions:
    get:
      tags:
      - Auction Management 
      summary: Returns all Auctions
      description: To get list of all Auction.
      operationId: getListOfAllAuction
      responses:
        200:
          description: "List of All Auction"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAuctionResponseBody"
        204:
          description: "No Content"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoContent"        
        400:
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequest"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unauthorized"
        403:
            description: "Forbidden"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Forbidden"
        404:
          description: "Page Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
      # security:
      # - AuctionAuthorizer: []
    
    post:
      tags:
      - Auction Management 
      summary: Create New Auction 
      description: To Create New Auction.
      operationId: addAuction
      requestBody:
        description: Model to post auction details
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AuctionRequestBody'     
      responses:
        200:
          description: "Auction Created With Following values"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionResponseBody"
        204:
          description: "No Content"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoContent"        
        400:
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequest"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unauthorized"
        403:
            description: "Forbidden"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Forbidden"
        404:
          description: "Page Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
      security:
      - AuctionAuthorizer: []
...
  /auctions/{auction_id}/players:
    post:
        tags:
        - Auction Player Management 
        summary: Add Selected Players In Auction 
        description: To add Selected Players in Auction.
        operationId: addSelectedPlayersInAuction
        parameters:
        - in: path
          name: auction_id
          required: true
          schema:
            type: integer
        requestBody:
          description: Model to post selected player details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelectAuctionPlayerRequestBody'     
        responses:
          200:
            description: "Selected Player Created In Auction With Following values"
            headers:
              Access-Control-Allow-Origin:
                schema:
                  type: "string"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/SelectAuctionPlayerResponseBody"
          204:
            description: "No Content"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NoContent"        
          400:
            description: "Bad Request"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/BadRequest"
          401:
            description: "Unauthorized"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Unauthorized"
          403:
            description: "Forbidden"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Forbidden"
          404:
            description: "Page Not Found"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NotFound"        
          500:
            description: "Internal Server Error"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/InternalServerError"
        security:
        - AuctionAuthorizer: []
...
  /auctions/{auction_id}/players/{player_id}:
    delete:
        tags:
        - Auction Player Management 
        summary: Delete Player In Auction 
        description: To delete a Player from Auction.
        operationId: deleteAuctionPlayer
        parameters:
        - in: path
          name: player_id
          required: true
          schema:
            type: integer
        - in: path
          name: auction_id
          required: true
          schema:
            type: integer
        responses:
          200:
            description: "Auction Player Deleted"
            headers:
              Access-Control-Allow-Origin:
                schema:
                  type: "string"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/AuctionPlayerDeleted"
          204:
            description: "No Content"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NoContent"        
          400:
            description: "Bad Request"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/BadRequest"
          401:
            description: "Unauthorized"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Unauthorized"
          403:
            description: "Forbidden"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Forbidden"
          404:
            description: "Page Not Found"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NotFound"        
          500:
            description: "Internal Server Error"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/InternalServerError"                
...
components:
  schemas:
...
    InternalServerError:
      title: Internal server error occurred
      type: object
      properties:
        message:
          type: string
          example: Internal Server Error      
    Unauthorized:
      title: Unauthorized
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
    Forbidden:
      title: Forbidden
      type: object
      properties:
        message:
          type: string
          example: Forbidden
    NoContent:
      title: NoContent
      type: object
      properties:
        message:
          type: string
          example: No Content
    BadRequest:
      title: Invalid input parameters
      type: object
      properties:
        message:
          type: string
          example: Bad Request
...
    AuctionRequestBody:
      type: object
      properties:
        auction_id:
          type: integer
        auction_name:
          type: string
        start_time:
          type: string
          example: "2020-10-04T17:42:49"
        end_time:
          type: string
          example: "2020-11-01T17:42:49"
        no_of_teams:
          type: string
        min_team_size:
          type: string
        max_team_size:
          type: string        
        purse_value:
          type: string
        is_deleted:
          type: boolean
        auction_country:
          type: string
          example: "India"
        auction_description:
          type: string
          example: "T20 Auction"
        auction_image:
          type: string
          format: binary
    ListAuctionResponseBody:
          type: object
          properties:
            auction_name:
              type: string
            start_time:
              type: string
              example: "2020-10-04T17:42:49"
            end_time:
              type: string
              example: "2020-11-01T17:42:49"
            no_of_teams:
              type: integer
            min_team_size:
              type: integer
            max_team_size:
              type: integer
            purse_value:
              type: integer 
    AuctionResponseBody:
      type: object
      properties:
        auction_id:
          type: integer
        auction_name:
          type: string
        no_of_teams:
          type: integer
        min_team_size:
          type: integer
        max_team_size:
          type: integer
        purse_value:
          type: integer
        is_deleted:
          type: boolean 
...
    AuctionPlayerDeleted:
      title: Successfully Deleted
      type: object
      properties:
        message:
          type: string
          example: Player Deleted Successfully 
...
    SelectAuctionPlayerRequestBody:
      type: object
      properties:
        selected_players:
          type: "array"
          items:
            type: "object"
            properties:
              player_id:
                type: integer
                example: 101
              base_value:
                type: integer
                example: 400000
    SelectAuctionPlayerResponseBody:
      type: "array"
      items:
        type: "object"
        properties:
          auction_id:
            type: integer
            example: 1
          player_id:
            type: integer
            example: 101
          team_id:
            type: integer
            example: 201
          sold_status:
            type: string
            example: "Pending"
          base_value:
            type: integer
            example: 400000
          sold_value:
            type: integer
            example: 800000
...
    NotFound:
      title: NotFound
      type: object
      properties:
        message:
          type: string
          example: Not Found
...
  securitySchemes:
    AuctionAuthorizer:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      x-amazon-apigateway-authtype: "custom"
      x-amazon-apigateway-authorizer:        
        todo-authorizerUri: 
          Fn::Sub: 'todo-auth'
        authorizerResultTtlInSeconds: 0
        identitySource: "method.request.header.Authorization"
        type: "request"